extmod: Move modnetwork and modusocket to extmod.
