+++ ports/stm32/modnwcc3k.c
@@ -37,7 +37,7 @@
 #include "py/mperrno.h"
 #include "py/mphal.h"
 #include "shared/netutils/netutils.h"
-#include "modnetwork.h"
+#include "extmod/modnetwork.h"
 #include "pin.h"
 #include "spi.h"
